pull up locator normalization to module
