+++ lib/locreator.js
@@ -1,31 +1,11 @@
 'use strict';
 var Drivex = require('selenium-drivex');
 var _ = require('lodash');
+var normalize = require('./normalize');
 
-var _normify = function (nemo, _locator) {
-  var locator = _locator;
-  if (_locator.constructor === String) {
-    locator = _splitLocator(_locator);
-  }
-  return nemo.wd.By[locator.type](locator.locator);
-};
-
-var _splitLocator = function (locatorString) {
-  var splitLocator = locatorString.split(':');
-  if (splitLocator.length === 1) {
-    splitLocator[1] = splitLocator[0];
-    splitLocator[0] = 'css';
-  }
-  var locator = {
-    'type': splitLocator[0],
-    'locator': splitLocator[1]
-  };
-  return locator;
-};
-module.exports.normify = _normify;
 module.exports.addGenericMethods = function generics(nemo) {
   var normify = function (locator) {
-    return _normify(nemo, locator);
+    return normalize(nemo, locator);
   };
   var drivex = Drivex(nemo.driver, nemo.wd);
 
@@ -59,7 +39,7 @@ module.exports.addGenericMethods = function generics(nemo) {
 module.exports.addStarMethods = function locreator(nemo, locatorId, _locator, parentWebElement) {
   var locatorObject = {};
   var drivex = Drivex(nemo.driver, nemo.wd);
-  var locator = _normify(nemo, _locator);
+  var locator = normalize(nemo, _locator);
 
   locatorObject[locatorId] = function () {
     return drivex.find(locator, parentWebElement);
