+++ lib/view.js
@@ -25,6 +25,7 @@
  instantiate any subviews
  */
 var locreator = require('./locreator');
+var normalize = require('./normalize');
 var Drivex = require('selenium-drivex');
 var Locatex = require('./locatex');
 var _ = require('lodash');
@@ -39,7 +40,7 @@ module.exports = function View(nemo, viewJSON) {
 
     if (viewJSON[locatorId].Elements) {
       viewObject[locatorId] = function () { //give back the nemo.view.viewname.list() function
-        return drivex.finds(locreator.normify(nemo, locator)).then(function (parentWebElements) {
+        return drivex.finds(normalize(nemo, locator)).then(function (parentWebElements) {
           return nemo.wd.promise.map(parentWebElements, function (parentWebElement) {
             var parentObject = {};
             Object.keys(viewJSON[locatorId].Elements).forEach(function (childLocatorId) {
